+++ Editor/NFTPortSettings.cs
@@ -2,6 +2,7 @@ using System;
 using UnityEngine;
 using UnityEditor;
 using Newtonsoft.Json;
+using UnityEditor.PackageManager;
 
 namespace NFTPort.Editor
 { using Internal;
@@ -9,12 +10,32 @@ namespace NFTPort.Editor
     public class NFTPortSettings : EditorWindow
     {
         
-        public static string myAPIString = "Enter Your API Key";
-    
+        static string myAPIString = PortConstants.DefaultAPIKey;
+         
+        protected static Type WindowType = typeof(NFTPortSettings);
+        private static bool windowopen = false;
+        private bool ranLatestrel = false;
+        PkgJson releasedPkgJson = null;
         [MenuItem("NFTPort/Home")]
         public static void ShowWindow()
         {
-            GetWindow<NFTPortSettings>("Home");
+            var win = GetWindow<NFTPortSettings>(PortConstants.HomeWindowName);
+            SetSize(win);
+        }
+        
+        /// <summary>   
+        /// When the package is imported
+        /// </summary>
+        [UnityEditor.InitializeOnLoadMethod]
+        public static void InitializeOnLoadMethod()
+        {
+            Events.registeredPackages += RegisteredPackagesEventHandler;
+            //EditorApplication.delayCall += delayCall;
+        }
+
+        static void RegisteredPackagesEventHandler(PackageRegistrationEventArgs packageRegistrationEventArgs)
+        {
+            ReadFromUserPrefs(); 
         }
 
         void OnGUI()
@@ -22,45 +43,148 @@ namespace NFTPort.Editor
             Texture banner = Resources.Load<Texture>("banner");
             GUILayout.Box(banner);
         
-            GUILayout.Label("Welcome to NFTPort Unity SDK ", EditorStyles.boldLabel);
+            GUILayout.Label("Welcome to NFTPort Unity SDK ", EditorStyles.whiteLargeLabel);
+            GUILayout.Label("\n" +
+                            " Create cross-chain compatible NFT games \n and products in Unity with fast and reliable data.\n" +
+                            "", EditorStyles.label);
 
-            myAPIString = EditorGUILayout.TextField("APIKEY", myAPIString);
             
-            if (GUILayout.Button("Save"))
+            GUILayout.BeginHorizontal("box");
+            var defaultColor = GUI.backgroundColor;
+            if(APIkeyOk())
+                GUI.color = UnityEngine.Color.green;
+            else
+            {
+                GUI.color = UnityEngine.Color.red;
+            }
+            myAPIString = EditorGUILayout.TextField("APIKEY", myAPIString);
+            GUI.color = defaultColor;
+            GUILayout.EndHorizontal();
+            
+            
+            if (GUILayout.Button("Save API Key", GUILayout.Height(25)))
                 SaveChanges();
+            
+            EditorGUILayout.LabelField("");
+            
+            GuiLine();
+            
+            if (GUILayout.Button("View Documentation", GUILayout.Height(25)))
+                Application.OpenURL(PortConstants.Docs_GettingStarted);
+            
+            if (GUILayout.Button("Community & Support", GUILayout.Height(25)))
+                Application.OpenURL(PortConstants.DiscordInvite);
+            
+            GuiLine();
+            
+            EditorGUILayout.LabelField("");
+
+            if (userModel != null)
+            {
+                EditorGUILayout.LabelField("   Welcome " + userModel.profile.name + "."); 
+            }
+            
+            EditorGUILayout.LabelField("");
+            
+            GuiLine();
+            
+            GUILayout.BeginHorizontal("box");
+            EditorGUILayout.LabelField("installed version: " + PkgInfo.GetPackageVer());
+
+            if (!ranLatestrel)
+            {
+                var ls = LatestRel.Initialize();
+                if (ls != null)
+                {
+                    ls.OnComplete(pkg => releasedPkgJson = pkg);
+                    ls.Run();
+                }
+
+                ranLatestrel = true;
+            }
+
+            if (releasedPkgJson != null)
+            {
+                EditorGUILayout.LabelField("Latest release version: " + releasedPkgJson.version);
+            }
+
+            GUILayout.EndHorizontal();
+            
+            
+            if (GUILayout.Button("star the github", GUILayout.Height(17), GUILayout.Width(155)))
+                Application.OpenURL(PortConstants.Github);
+
         }
 
         void OnEnable()
         {
-            ReadFromUserPrefs();
+            if(!windowopen)
+                ReadFromUserPrefs();
+            windowopen = true;
+            ranLatestrel = false;
+        }
+
+        private void OnDisable()
+        {
+            windowopen = false;
         }
 
         public override void SaveChanges()
         {
             WriteToUserPrefs();
+            UserStats(); 
         }
         
-        
-        
+        static void ShowHomeWindow()
+        {
+            if(myAPIString != PortConstants.DefaultAPIKey || windowopen)
+                return;
+            
+            NFTPortSettings win = GetWindow(WindowType, false, PortConstants.HomeWindowName, true) as NFTPortSettings;
+            if (win == null)
+            {
+                return;  
+            }
 
+            windowopen = true;
+            SetSize(win);
+            win.Show();
+        }
+        static void SetSize(NFTPortSettings win) 
+        {
+            win.minSize = new Vector2(530, 600);
+            win.maxSize = new Vector2(530, 600);
+        } 
+        
+        static void GuiLine( int i_height = 1 )
+        {
+            Rect rect = EditorGUILayout.GetControlRect(false, i_height );
+            rect.height = i_height;
+            EditorGUI.DrawRect(rect, new UnityEngine.Color ( 0.5f,0.5f,0.5f, 1 ) );
+        }
+        
         #region ReadWrite UserPrefs
         
-        [SerializeField]
-        private PortUser.UserPrefs _userPrefs = new PortUser.UserPrefs();
-        void ReadFromUserPrefs()
+        private static PortUser.UserPrefs _userPrefs = new PortUser.UserPrefs();
+        private static TextAsset targetFile;
+        static void ReadFromUserPrefs()
         {
-            TextAsset targetFile = Resources.Load<TextAsset>("NFTPort UserPrefs");;
+            targetFile = Resources.Load<TextAsset>("NFTPort UserPrefs");
             if (targetFile != null)
             {
                 _userPrefs = JsonConvert.DeserializeObject<PortUser.UserPrefs>(targetFile.text);
                 myAPIString = _userPrefs.API_KEY;
+                //PortUser.Initialise();
+                UserStats();
             }
             else
-            {
+            { 
                 PortUser._initialised = false;
-                myAPIString = "Enter your API KEY here";
+                myAPIString = PortConstants.DefaultAPIKey;
+                if(!windowopen)
+                    ShowHomeWindow();
             }
-                
+
         }
         void WriteToUserPrefs()
         {
@@ -69,8 +193,41 @@ namespace NFTPort.Editor
         }
 
         #endregion
-     
-        
+
+        #region Userstats
+        static User_model userModel;
+        static void UserStats()
+        {
+            userModel = null;
+            User_Settings
+                .Initialize(true)
+                .OnComplete(usermodel=> userModel = usermodel)
+                .Run();
+        }
+         
+
+        bool APIkeyOk()
+        {
+            if (userModel == null)
+            {
+                PortUser._initialised = false;
+                return false;
+            }
+
+            if (userModel.response == "OK")
+            {
+                PortUser.Initialise();
+                return true;
+            }
+            else
+            {
+                PortUser._initialised = false;
+                return false;
+            }
+        }
+
+        #endregion
+         
         
     }
 }