+++ Runtime/Internal/PortConstants.cs
@@ -1,0 +1,14 @@
+namespace NFTPort.Internal
+{
+    public static class PortConstants
+    {
+        public const string HomeWindowName = "NFTPort Home";
+        public const string DefaultAPIKey = "Enter your API KEY here";
+
+        public const string Docs_GettingStarted = "https://docs.nftport.xyz/docs/nftport/ZG9jOjUwMzc2NzA5-getting-started";
+        public const string DiscordInvite = "https://discord.gg/w92sXkNmBR";
+        public const string Github = "https://github.com/nftport/nftport-unity.git";
+
+    }
+}
+